[21004] Consolidating MPI User Identity and Relationship functions
