+++ lib/mpi/models/mvi_profile.rb
@@ -2,57 +2,19 @@
 
 require 'common/models/attribute_types/date_time_string'
 require_relative 'mvi_profile_address'
+require_relative 'mvi_profile_identity'
+require_relative 'mvi_profile_ids'
+require_relative 'mvi_profile_relationship'
 
 module MPI
   module Models
     class MviProfile
       include Virtus.model
+      include MviProfileIdentity
+      include MviProfileIds
 
-      attribute :given_names, Array[String]
-      attribute :family_name, String
-      attribute :suffix, String
-      attribute :gender, String
-      attribute :birth_date, Common::DateTimeString
-      attribute :ssn, String
-      attribute :address, MviProfileAddress
-      attribute :home_phone, String
-      attribute :full_mvi_ids, Array[String]
-      attribute :icn, String
-      attribute :icn_with_aaid, String
-      attribute :mhv_ids, Array[String]
-      attribute :active_mhv_ids, Array[String]
-      attribute :vha_facility_ids, Array[String]
-      attribute :edipi, String
-      attribute :participant_id, String
-      attribute :birls_id, String
-      attribute :birls_ids, Array[String]
-      attribute :sec_id, String
-      attribute :vet360_id, String
-      attribute :historical_icns, Array[String]
       attribute :search_token, String
-      attribute :cerner_facility_ids, Array[String]
-      attribute :cerner_id, String
-      attribute :person_type_code, String
-      attribute :relationships, Array[MviProfile]
-
-      def mhv_correlation_id
-        @active_mhv_ids&.first
-      end
-
-      def normalized_suffix
-        case @suffix
-        when /jr\.?/i
-          'Jr.'
-        when /sr\.?/i
-          'Sr.'
-        when /iii/i
-          'III'
-        when /ii/i
-          'II'
-        when /iv/i
-          'IV'
-        end
-      end
+      attribute :relationships, Array[MviProfileRelationship]
     end
   end
 end
