+++ lib/mpi/responses/profile_parser.rb
@@ -78,21 +78,52 @@ module MPI
 
       private
 
-      # rubocop:disable Metrics/MethodLength
       def build_mvi_profile(patient)
-        name = parse_name(locate_element(patient, PATIENT_PERSON_PREFIX + NAME_XPATH))
-        full_mvi_ids = get_extensions(patient.locate('id'))
-        parsed_mvi_ids = parse_xml_gcids(patient.locate('id'))
-        log_inactive_mhv_ids(parsed_mvi_ids[:mhv_ids].to_a, parsed_mvi_ids[:active_mhv_ids].to_a)
-        MPI::Models::MviProfile.new(
+        historical_icns = @original_body.locate(HISTORICAL_ICN_XPATH)
+        profile_identity_hash = create_mvi_profile_identity(locate_element(patient, PATIENT_PERSON_PREFIX))
+        profile_ids_hash = create_mvi_profile_ids(patient, historical_icns)
+        misc_hash = {
+          search_token: locate_element(@original_body, 'id').attributes[:extension],
+          relationships: parse_relationships(patient.locate(PATIENT_RELATIONSHIP_XPATH))
+        }
+
+        MPI::Models::MviProfile.new(profile_identity_hash.merge(profile_ids_hash).merge(misc_hash))
+      end
+
+      def parse_relationships(relationships_array)
+        relationships_array.map { |relationship| build_relationship_mvi_profile(relationship) }
+      end
+
+      def build_relationship_mvi_profile(relationship)
+        relationship_component = locate_element(relationship, RELATIONSHIP_PREFIX)
+        relationship_identity_hash = create_mvi_profile_identity(relationship_component)
+        relationship_ids_hash = create_mvi_profile_ids(relationship_component)
+
+        MPI::Models::MviProfileRelationship.new(relationship_identity_hash.merge(relationship_ids_hash))
+      end
+
+      def create_mvi_profile_identity(person_component)
+        name = parse_name(locate_element(person_component, NAME_XPATH))
+
+        {
           given_names: name[:given],
           family_name: name[:family],
           suffix: name[:suffix],
-          gender: locate_element(patient, PATIENT_PERSON_PREFIX + GENDER_XPATH),
-          birth_date: locate_element(patient, PATIENT_PERSON_PREFIX + DOB_XPATH),
-          ssn: parse_ssn(locate_element(patient, PATIENT_PERSON_PREFIX + SSN_XPATH)),
-          address: parse_address(patient),
-          home_phone: parse_phone(patient),
+          gender: locate_element(person_component, GENDER_XPATH),
+          birth_date: locate_element(person_component, DOB_XPATH),
+          ssn: parse_ssn(locate_element(person_component, SSN_XPATH)),
+          address: parse_address(person_component),
+          home_phone: parse_phone(person_component),
+          person_type_code: nil
+        }
+      end
+
+      def create_mvi_profile_ids(patient, historical_icns = nil)
+        full_mvi_ids = get_extensions(patient.locate('id'))
+        parsed_mvi_ids = parse_xml_gcids(patient.locate('id'))
+        log_inactive_mhv_ids(parsed_mvi_ids[:mhv_ids].to_a, parsed_mvi_ids[:active_mhv_ids].to_a)
+
+        {
           full_mvi_ids: full_mvi_ids,
           icn: parsed_mvi_ids[:icn],
           mhv_ids: parsed_mvi_ids[:mhv_ids],
@@ -104,15 +135,12 @@ module MPI
           birls_id: sanitize_id(parsed_mvi_ids[:birls_id]),
           birls_ids: sanitize_id_array(parsed_mvi_ids[:birls_ids]),
           vet360_id: parsed_mvi_ids[:vet360_id],
-          historical_icns: parse_xml_historical_icns(@original_body.locate(HISTORICAL_ICN_XPATH)),
+          historical_icns: parse_xml_historical_icns(historical_icns),
           icn_with_aaid: parsed_mvi_ids[:icn_with_aaid],
-          search_token: locate_element(@original_body, 'id').attributes[:extension],
           cerner_id: parsed_mvi_ids[:cerner_id],
-          cerner_facility_ids: parsed_mvi_ids[:cerner_facility_ids],
-          relationships: parse_relationships(patient.locate(PATIENT_RELATIONSHIP_XPATH))
-        )
+          cerner_facility_ids: parsed_mvi_ids[:cerner_facility_ids]
+        }
       end
-      # rubocop:enable Metrics/MethodLength
 
       def get_extensions(id_array)
         id_array.map do |id_object|
@@ -136,48 +164,6 @@ module MPI
                                 ids: active_mhv_ids)
         end
       end
-
-      def parse_relationships(relationships_array)
-        relationships_array.map { |relationship| build_relationship_mvi_profile(relationship) }
-      end
-
-      # rubocop:disable Metrics/MethodLength
-      # rubocop:disable Metrics/AbcSize
-      def build_relationship_mvi_profile(relationship)
-        relationship_holder = locate_element(relationship, RELATIONSHIP_PREFIX)
-        name = parse_name(locate_element(relationship_holder, NAME_XPATH))
-        full_mvi_ids = get_extensions(relationship_holder.locate('id'))
-        parsed_mvi_ids = parse_xml_gcids(relationship_holder.locate('id'))
-        log_inactive_mhv_ids(parsed_mvi_ids[:mhv_ids].to_a, parsed_mvi_ids[:active_mhv_ids].to_a)
-        MPI::Models::MviProfile.new(
-          person_type_code: locate_element(relationship, 'code').attributes[:code],
-          given_names: name[:given],
-          family_name: name[:family],
-          suffix: name[:suffix],
-          gender: locate_element(relationship_holder, GENDER_XPATH),
-          birth_date: locate_element(relationship_holder, DOB_XPATH),
-          ssn: parse_ssn(locate_element(relationship_holder, SSN_XPATH)),
-          address: nil,
-          home_phone: locate_element(relationship, PHONE).attributes[:value],
-          full_mvi_ids: full_mvi_ids,
-          icn: parsed_mvi_ids[:icn],
-          mhv_ids: parsed_mvi_ids[:mhv_ids],
-          active_mhv_ids: parsed_mvi_ids[:active_mhv_ids],
-          edipi: sanitize_edipi(parsed_mvi_ids[:edipi]),
-          participant_id: sanitize_id(parsed_mvi_ids[:vba_corp_id]),
-          vha_facility_ids: parsed_mvi_ids[:vha_facility_ids],
-          sec_id: parsed_mvi_ids[:sec_id],
-          birls_id: sanitize_id(parsed_mvi_ids[:birls_id]),
-          birls_ids: parsed_mvi_ids[:birls_ids].map { |id| sanitize_id(id) }.compact,
-          vet360_id: parsed_mvi_ids[:vet360_id],
-          # historical_icns: MPI::Responses::HistoricalIcnParser.new(@original_body).get_icns,
-          icn_with_aaid: parsed_mvi_ids[:icn_with_aaid],
-          cerner_id: parsed_mvi_ids[:cerner_id],
-          cerner_facility_ids: parsed_mvi_ids[:cerner_facility_ids]
-        )
-      end
-      # rubocop:enable Metrics/AbcSize
-      # rubocop:enable Metrics/MethodLength
 
       # name can be a hash or an array of hashes with extra unneeded details
       # given may be an array if it includes middle name
@@ -205,8 +191,8 @@ module MPI
         nil
       end
 
-      def parse_address(patient)
-        el = locate_element(patient, PATIENT_PERSON_PREFIX + ADDRESS_XPATH)
+      def parse_address(person)
+        el = locate_element(person, ADDRESS_XPATH)
         return nil unless el
 
         address_hash = el.nodes.map { |n| { n.value.snakecase.to_sym => n.nodes.first } }.reduce({}, :merge)
@@ -214,8 +200,8 @@ module MPI
         MPI::Models::MviProfileAddress.new(address_hash)
       end
 
-      def parse_phone(patient)
-        el = locate_element(patient, PATIENT_PERSON_PREFIX + PHONE)
+      def parse_phone(person)
+        el = locate_element(person, PHONE)
         return nil unless el
 
         el.attributes[:value]
